requirements.txt
python-box
protobuf_to_dict
shapely
